+++ test/acceptance/docker-compose.debug.yaml
@@ -1,4 +1,4 @@
 version: '3.7'
 services:
   wdio-cucumber:
-    command: npm run test:debug
+    command: npm run test
