+++ test/acceptance/docker-compose.override.yaml
@@ -14,6 +14,7 @@ services:
       - ./wdio.conf:/home/node/wdio.conf
       - ./package.json:/home/node/package.json
       - ./html-reports/:/home/node/html-reports/
+      
     ports:
       - "9228:9229"
 
@@ -28,3 +29,10 @@ services:
   hub:
     ports:
       - "4444:4444"
+
+  browserstack:
+    ports:
+      - "45454:45454"
+   
+
+
