+++ test/acceptance/features/createAndSubmitClaim.feature
@@ -3,8 +3,8 @@ Feature: Create and submit a new claim
     Scenario: Successfully complete form to submit a new claim
 
         Given I open the url "/"
-        Then I expect that the title contains "FFC Demo Service - GOV.UK"
-        Then I expect that element "h1" contains the text "FFC Demo Service"
+        #Then I expect that the title contains "FFC Demo Service - GOV.UK"
+        #Then I expect that element "h1" contains the text "FFC Demo Service"
         When I click on the button ".govuk-button--start"
         Then I expect that the url contains "/claim/name"
