+++ test/acceptance/wdio.conf.js
@@ -1,4 +1,8 @@
 var browserstack = require('browserstack-local')
+// var request = require('request')
+// const compatibility = require('./browsers')
+// const moment = require('moment-timezone')
+// const timestamp = moment.tz('Europe/London').format('D/M/YY hh:mm:ss')
 const { ReportAggregator, HtmlReporter } = require('@rpii/wdio-html-reporter')
 const log4js = require('@log4js-node/log4js-api')
 const logger = log4js.getLogger('default')
@@ -7,23 +11,55 @@ const chromeArgs = process.env.CHROME_ARGS ? process.env.CHROME_ARGS.split(' ') 
 const maxInstances = process.env.MAX_INSTANCES ? Number(process.env.MAX_INSTANCES) : 5
 
 exports.config = {
+  // hostname: 'selenium',  //
+  // hostname: 'hub-cloud.browserstack.com',
+  // path: '/wd/hub',
+  user: process.env.BROWSERSTACK_USERNAME,
+  key: process.env.BROWSERSTACK_ACCESS_KEY,
 
-  user: process.env.BROWSERSTACK_USERNAME || 'kaziyiola2',
-  key: process.env.BROWSERSTACK_ACCESS_KEY || '9wMS2xKkzpvkwsrcs9Vt',
-  hostname: 'selenium',
-  path: '/wd/hub',
   specs: ['./features/**/*.feature'],
   exclude: ['./scratch/**'],
+
+  // capabilities: compatibility.map(e => {
+  //   e.project = 'ffc-demo-web'
+  //   e.build = 'ffc-demo-web'
+  //   e.name = `Kaz iyiola@ ${timestamp}`
+  //   e['browserstack.local'] = true
+  //   e['browserstack.debug'] = true
+  //   e['browserstack.networkLogs'] = true
+  //   e['browserstack.acceptSslCerts'] = true
+  //   e['browserstack.console'] = 'errors'
+  //   e['browserstack.use_w3c'] = true
+  //   e['browserstack.selenium_version'] = '3.141.59'
+  //   return e
+  // }),
+
   maxInstances,
   capabilities: [{
     maxInstances,
     acceptInsecureCerts: true,
     browserName: 'chrome',
     'browserstack.local': true,
+    'browserstack.networkLogs': true,
+    'browserstack.acceptSslCerts': true,
     'goog:chromeOptions': {
       args: chromeArgs
     }
-  }],
+  },
+  {
+    // Windows 10 Firefox latest
+    os: 'Windows',
+    osVersion: '10',
+    browserName: 'Firefox',
+    browserVersion: 'latest'
+  },
+  {
+    // MacOS Mojave Safari 13.1
+    os: 'OS X',
+    osVersion: 'Catalina',
+    browserName: 'Safari',
+    browserVersion: 'latest'
+  } ],
   //
   // ===================
   // Test Configurations
@@ -31,7 +67,8 @@ exports.config = {
   // Define all options that are relevant for the WebdriverIO instance here
   logLevel: 'warn',
   bail: 0,
-  baseUrl: envRoot + '/selenium:4444',
+  // baseUrl: envRoot + '/selenium:4444',
+  baseUrl: envRoot,
   waitforTimeout: 10000,
   connectionRetryTimeout: 90000,
   connectionRetryCount: 3,
@@ -86,42 +123,31 @@ exports.config = {
     })
     reportAggregator.clean()
     global.reportAggregator = reportAggregator
-    // starting  browser stack
-    console.log(process.env.BROWSERSTACK_ACCESS_KEY)
+
     console.log('Connecting local')
     return new Promise(function (resolve, reject) {
       exports.bs_local = new browserstack.Local()
-      exports.bs_local.start({ 'key': exports.config.key }, function (error) {
+      const bsLocalArgs = {
+        key: exports.config.key,
+        verbose: 'true',
+        force: 'true',
+        onlyAutomate: 'true',
+        forceLocal: 'true'
+      }
+      exports.bs_local.start(bsLocalArgs, function (error) {
         if (error) return reject(error)
-
         console.log('Connected. Now testing...')
         resolve()
       })
-    }) // end of starting browser stack
+    })
   },
-
-  // end of starting browser stack
-
-  // Code to mark the status of test on BrowserStack based on the assertion status
-  //   afterTest: function (test, context, { error, result, duration, passed, retries }) {
-  //     if (passed) {
-  //       browser.executeScript('browserstack_executor: {"action": "setSessionStatus", "arguments": {"status":"passed","reason": "Assertions passed"}}')
-  //     } else {
-  //       browser.executeScript('browserstack_executor: {"action": "setSessionStatus", "arguments": {"status":"failed","reason": "At least 1 assertion failed"}}')
-  //     }
-  //   },
 
   onComplete: function (exitCode, config, capabilities, results) {
     (async () => {
       await global.reportAggregator.createReport()
     })()
-    // stop browser test
-    return new Promise(function (resolve, reject) {
-      exports.bs_local.stop(function () {
-        console.log('Binary stopped')
-        resolve()
-      })
-    })
+    exports.bs_local.stop()
+    console.log('Testing complete, binary closed')
   },
 
   beforeSession: function () {
